Refactored/moved TokenID(...) and adapted skaffold dependency configuration
